Refactor cache into seperate files
